refactor and rename handler.ts
